Arc length extrusion

+++ tests/gcode/MockPrinter.py
@@ -33,6 +33,7 @@ from CascadingConfigParser import CascadingConfigParser
 from Redeem import *
 from EndStop import EndStop
 
+
 """
 Override CascadingConfigParser methods to set self. variables
 """
@@ -41,18 +42,46 @@ class CascadingConfigParserWedge(CascadingConfigParser):
         self.replicape_revision = "0A4A" # Fake. No hardware involved in these tests (Redundant?)
         self.reach_revision = "00A0" # Fake. No hardware involved in these tests (Redundant?)
 
-"""
-MockPrinter, in combination with the many sys.module[...] = Mock() statements
-above, creates a mock Redeem instance. The mock instance has only what is
-needed for our tests and does not access any BBB hardware IOs.
-"""
+
 class MockPrinter(unittest.TestCase):
+
+    """
+    MockPrinter, in combination with the many sys.module[...] = Mock() statements
+    above, creates a mock Redeem instance. The mock instance has only what is
+    needed for our tests and does not access any BBB hardware IOs.
+    """
+
+    @classmethod
+    def setUpPatch(cls):
+        """"
+        Override this method for mocking something other than the path planner
+        """
+        cls.printer.path_planner = mock.MagicMock()
+
+    @classmethod
+    def setUpConfigFiles(cls, path):
+        """
+        This seems like the best way to add to or change stuff in default.cfg,
+        without actually messing with the prestine file. Overwrite if you want
+        different printer.cfg and/or local.cfg files. For example, copy example filles...
+
+        copyfile(os.path.join(os.path.dirname(__file__), "my_test_local.cfg"), os.path.join(path, 'local.cfg'))
+        copyfile(os.path.join(os.path.dirname(__file__), "my_test_printer.cfg"), os.path.join(path, 'printer.cfg'))
+
+        """
+        tf = open("../configs/local.cfg", "w")
+        lines = """
+        [System]
+        log_to_file = False
+        """
+        tf.write(lines)
+        tf.close()
 
     @classmethod
     @mock.patch.object(EndStop, "_wait_for_event", new=None)
     @mock.patch.object(CascadingConfigParser, "get_key")
-    @mock.patch("Redeem.CascadingConfigParser", new = CascadingConfigParserWedge)
-    def setUpClass(self, mock_get_key):
+    @mock.patch("Redeem.CascadingConfigParser", new=CascadingConfigParserWedge)
+    def setUpClass(cls, mock_get_key):
 
         mock_get_key.return_value = "TESTING_DUMMY_KEY"
 
@@ -61,6 +90,8 @@ class MockPrinter(unittest.TestCase):
         """
         class DisabledExtruder(Extruder):
             def enable(self):
+                self.avg = 1
+                self.temperatures = [100]
                 pass
         class DisabledHBP(HBP):
             def enable(self):
@@ -68,57 +99,56 @@ class MockPrinter(unittest.TestCase):
         mock.patch('Redeem.Extruder', side_effect=DisabledExtruder).start()
         mock.patch('Redeem.HBP', side_effect=DisabledHBP).start()
 
+        cfg_path = "../configs"
+        cls.setUpConfigFiles(cfg_path)
 
+        cls.R = Redeem(config_location=cfg_path)
+        cls.printer = cls.R.printer
 
-        """
-        This seems like the best way to add to or change stuff in default.cfg,
-        without actually messing with the prestine file.
-        """
-        tmp_local_cfg = tf = open("../configs/local.cfg", "w")
-        lines = """
-[System]
-log_to_file = False
-"""
-        tf.write(lines)
-        tf.close()
+        cls.setUpPatch()
 
-        self.R = Redeem(config_location="../configs")
-        printer = self.printer = self.R.printer
+        cls.gcodes = cls.printer.processor.gcodes
+        cls.printer.send_message = mock.create_autospec(cls.printer.send_message)
 
-        self.printer.path_planner = mock.MagicMock()
-        self.gcodes = self.printer.processor.gcodes
-        self.printer.send_message = mock.create_autospec(self.printer.send_message)
+        cls.printer.movement = Path.ABSOLUTE
+        cls.printer.feed_rate = 0.050  # m/s
+        cls.printer.accel = 0.050 / 60  # m/s/s
 
-        self.printer.movement = Path.ABSOLUTE
-        self.printer.feed_rate = 0.050 # m/s
-        self.printer.accel = 0.050 / 60 # m/s/s
+        Gcode.printer = cls.printer
+        Path.printer = cls.printer
 
-        Gcode.printer = printer
-        Path.printer = printer
-
-        self.printer.speed_factor = 1.0
+        cls.printer.speed_factor = 1.0
         """ 
         We want to ensure that printer.factor is always obeyed correctly
         For convenience, we'll set it to mm/inch and check that resulting 
         paths have the correct meter values, converted from inch input.
         """
-        self.printer.unit_factor = self.f = 25.4 # inches
+        cls.printer.unit_factor = cls.f = 25.4  # inches
 
-        self.printer.probe_points = []
+        cls.printer.probe_points = []
 
     @classmethod
-    def tearDownClass(self):
-        self.R = self.printer = None
-        os.remove("../configs/local.cfg")
-        pass
+    def tearDownClass(cls):
+        cls.R = cls.printer = None
+        if os.path.exists("../configs/local.cfg"):
+            os.remove("../configs/local.cfg")
+        if os.path.exists("../configs/printer.cfg"):
+            os.remove("../configs/printer.cfg")
 
     """ directly calls a Gcode class's execute method, bypassing printer.processor.execute """
     @classmethod
-    def execute_gcode(self, text):
+    def execute_gcode(cls, text):
         g = Gcode({"message": text})
-        self.printer.processor.gcodes[g.gcode].execute(g)
+        cls.printer.processor.gcodes[g.gcode].execute(g)
         return g
 
     @classmethod
-    def full_path(self, o):
-      return o.__module__ + "." + o.__class__.__name__
+    def full_path(cls, o):
+        return o.__module__ + "." + o.__class__.__name__
+
+    def assertCloseTo(self, a, b, msg=''):
+        """test to the nearest thousandth of a millimeter"""
+        match = np.isclose(a, b, rtol=1e-6, atol=1e-6)
+        if not match:
+            exception = self._formatMessage(msg, "{} is not close to {}: {}".format(a, b, msg))
+            raise self.failureException(exception)
